+++ packages/fleather/test/widgets/editor_test.dart
@@ -1,4 +1,6 @@
 import 'package:fleather/fleather.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -20,7 +22,8 @@ void main() {
       final BuildContext context = tester.element(find.byType(Container));
       var theme = FleatherThemeData.fallback(context)
           .copyWith(link: const TextStyle(color: Colors.red));
-      var editor = EditorSandBox(tester: tester, document: doc, theme: theme);
+      var editor =
+          EditorSandBox(tester: tester, document: doc, fleatherTheme: theme);
       await editor.pumpAndTap();
       // await tester.pumpAndSettle();
       final p = tester.widget(find.byType(RichText).first) as RichText;
@@ -28,8 +31,30 @@ void main() {
       expect(text.children!.first.style!.color, Colors.red);
     });
 
-    testWidgets('collapses selection when unfocused', (tester) async {
+    testWidgets('Hides toolbar and selection handles when text changed',
+        (tester) async {
+      const delta = TextEditingDeltaInsertion(
+        oldText: 'Add ',
+        textInserted: 'Test',
+        insertionOffset: 0,
+        selection: TextSelection.collapsed(offset: 0),
+        composing: TextRange.empty,
+      );
       final editor = EditorSandBox(tester: tester);
+      await editor.pump();
+      await tester.longPressAt(const Offset(20, 20));
+      await tester.pump();
+      expect(editor.findSelectionHandles(), findsNWidgets(2));
+      expect(find.byType(AdaptiveTextSelectionToolbar), findsOneWidget);
+      final state = tester.state(find.byType(RawEditor)) as RawEditorState;
+      state.updateEditingValueWithDeltas([delta]);
+      await tester.pump(throttleDuration);
+      expect(editor.findSelectionHandles(), findsNothing);
+      expect(find.byType(AdaptiveTextSelectionToolbar), findsNothing);
+    });
+
+    testWidgets('collapses selection when unfocused', (tester) async {
+      final editor = EditorSandBox(tester: tester, autofocus: true);
       await editor.pumpAndTap();
       await editor.updateSelection(base: 0, extent: 3);
       // expect(editor.findSelectionHandle(), findsNWidgets(2));
@@ -64,7 +89,8 @@ void main() {
       final finder = find.text('Paste');
       expect(finder, findsOneWidget);
       await tester.tap(finder);
-      await tester.pumpAndSettle();
+      // account for throttling of history update
+      await tester.pumpAndSettle(throttleDuration);
       expect(editor.document.toPlainText(), '$clipboardText\n');
     });
 
@@ -82,7 +108,8 @@ void main() {
       final finder = find.text('Paste');
       expect(finder, findsOneWidget);
       await tester.tap(finder);
-      await tester.pumpAndSettle();
+      // account for throttling of history update
+      await tester.pumpAndSettle(throttleDuration);
       expect(editor.document.toPlainText(), '$clipboardText\n');
     });
 
@@ -92,6 +119,148 @@ void main() {
           MaterialApp(home: FleatherEditor(controller: FleatherController()));
       await tester.pumpWidget(widget);
       // Fails if thrown
+    });
+
+    group('Text selection', () {
+      testWidgets('Can select last separated character in paragraph on iOS',
+          (tester) async {
+        const text = 'Test.';
+        final document = ParchmentDocument.fromJson([
+          {'insert': '$text\n'}
+        ]);
+        final editor = EditorSandBox(
+          tester: tester,
+          document: document,
+          autofocus: true,
+          theme: ThemeData(platform: TargetPlatform.iOS),
+        );
+        await editor.pump();
+        await tester.tapAt(tester.getBottomRight(find.byType(FleatherEditor)) -
+            const Offset(1, 1));
+        await tester.pump();
+        expect(
+            editor.selection,
+            const TextSelection.collapsed(
+                offset: text.length, affinity: TextAffinity.upstream));
+      });
+
+      testWidgets(
+          'Tapping after the beginning of a word moves cursor after word on iOS',
+          (tester) async {
+        final editor = EditorSandBox(
+          tester: tester,
+          autofocus: true,
+          theme: ThemeData(platform: TargetPlatform.iOS),
+        );
+        await editor.pump();
+        await tester.tapAt(tester.getBottomLeft(find.byType(FleatherEditor)) +
+            const Offset(10, -1));
+        await tester.pump();
+        expect(
+            editor.selection,
+            const TextSelection.collapsed(
+                offset: 4, affinity: TextAffinity.upstream));
+      });
+
+      testWidgets(
+          'Tapping before the beginning of a word moves cursor at the end of previous word on iOS',
+          (tester) async {
+        final document = ParchmentDocument.fromJson([
+          {'insert': 'ab cd ef\n'}
+        ]);
+        final editor = EditorSandBox(
+          tester: tester,
+          document: document,
+          autofocus: true,
+          theme: ThemeData(platform: TargetPlatform.iOS),
+        );
+        await editor.pump();
+        await tester.tapAt(tester.getBottomLeft(find.byType(FleatherEditor)) +
+            const Offset(48, -1));
+        await tester.pump();
+        expect(
+            editor.selection,
+            const TextSelection.collapsed(
+                offset: 3, affinity: TextAffinity.downstream));
+      });
+
+      testWidgets(
+          'Tapping moves the cursor right where user tapped on other platforms',
+          (tester) async {
+        final document = ParchmentDocument.fromJson([
+          {'insert': 'Test\n'}
+        ]);
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        await tester.tapAt(tester.getBottomLeft(find.byType(FleatherEditor)) +
+            const Offset(10, -1));
+        await tester.pump();
+        expect(
+            editor.selection,
+            const TextSelection.collapsed(
+                offset: 1, affinity: TextAffinity.upstream));
+      });
+
+      testWidgets('selection handles for iOS', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+        final document = ParchmentDocument();
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        final rawEditor = tester.widget<RawEditor>(find.byType(RawEditor));
+        expect(rawEditor.selectionControls,
+            const TypeMatcher<CupertinoTextSelectionControls>());
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('selection handles for macOS', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.macOS;
+        final document = ParchmentDocument();
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        final rawEditor = tester.widget<RawEditor>(find.byType(RawEditor));
+        expect(rawEditor.selectionControls,
+            const TypeMatcher<CupertinoDesktopTextSelectionControls>());
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('selection handles for Android', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.android;
+        final document = ParchmentDocument();
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        final rawEditor = tester.widget<RawEditor>(find.byType(RawEditor));
+        expect(rawEditor.selectionControls,
+            const TypeMatcher<MaterialTextSelectionControls>());
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('selection handles for Windows', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.windows;
+        final document = ParchmentDocument();
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        final rawEditor = tester.widget<RawEditor>(find.byType(RawEditor));
+        expect(rawEditor.selectionControls,
+            const TypeMatcher<DesktopTextSelectionControls>());
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('selection handles for Linux', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.linux;
+        final document = ParchmentDocument();
+        final editor =
+            EditorSandBox(tester: tester, document: document, autofocus: true);
+        await editor.pump();
+        final rawEditor = tester.widget<RawEditor>(find.byType(RawEditor));
+        expect(rawEditor.selectionControls,
+            const TypeMatcher<DesktopTextSelectionControls>());
+        debugDefaultTargetPlatformOverride = null;
+      });
     });
 
     group('didUpdateWidget', () {
@@ -148,43 +317,6 @@ void main() {
       });
     });
   });
-}
-
-class TestUpdateWidget extends StatefulWidget {
-  const TestUpdateWidget(
-      {Key? key, required this.focusNodeAfterChange, this.testField = false})
-      : super(key: key);
-
-  final FocusNode focusNodeAfterChange;
-  final bool testField;
-
-  @override
-  State<StatefulWidget> createState() => TestUpdateWidgetState();
-}
-
-class TestUpdateWidgetState extends State<TestUpdateWidget> {
-  FocusNode? focusNode;
-
-  @override
-  Widget build(BuildContext context) => Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          TextButton(
-            onPressed: () =>
-                setState(() => focusNode = widget.focusNodeAfterChange),
-            child: const Text('Change state'),
-          ),
-          widget.testField
-              ? FleatherField(
-                  controller: FleatherController(),
-                  focusNode: focusNode,
-                )
-              : FleatherEditor(
-                  controller: FleatherController(),
-                  focusNode: focusNode,
-                ),
-        ],
-      );
 }
 
 const clipboardText = 'copied text';