+++ packages/parchment/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.6.0
+
+* Allow to undo/redo changes from toolbar
+
+## 1.5.0
+
+* Format link after hitting newline
+
 ## 1.4.0
 
 * Support modifying document rules